refactor(services): migrate task domain to services/task/ subpackage
Migrate 7 task-related service files from flat modules to a
`services/task/` subpackage, following the lazy-import pattern
established by `services/issue/` and `services/pr/`.

Changes:
- Create `services/task/` subpackage with 5 modules:
  - classifier.py (TaskStatusBucket, classify_task_status)
  - show.py (TaskShowService, TaskShowResult, etc.)
  - status.py (TaskStatusData, fetch_task_status_data, etc.)
  - service.py (TaskService)
  - resume.py (merged TaskResumeUsecase/Candidates/Operations)

- Update `services/__init__.py` to import from new locations
- Update all external importers (7 source files, 16 test files)
- Update test patch/monkeypatch targets to new module paths

The merge of task_resume_*.py into task/resume.py reduces
cross-module dependencies while preserving all functionality.

All tests pass: pytest tests/vibe3/services/ tests/vibe3/commands/test_task*.py
Type checking: mypy src/vibe3/services/task/ ✓
Linting: ruff check src/vibe3/services/task/ ✓

Issue: #2102
