+++ src/vibe3/services/__init__.py
@@ -217,11 +217,11 @@ _SYMBOL_MODULES = {
     "SignatureService": "vibe3.services.signature_service",
     "SpecRefService": "vibe3.services.spec_ref_service",
     "StatusQueryService": "vibe3.services.status_query_service",
-    "TaskResumeOperations": "vibe3.services.task_resume_operations",
-    "TaskResumeUsecase": "vibe3.services.task_resume_usecase",
-    "TaskService": "vibe3.services.task_service",
-    "TaskShowResult": "vibe3.services.task_show_service",
-    "TaskStatusBucket": "vibe3.services.task_status_classifier",
+    "TaskResumeOperations": "vibe3.services.task.resume",
+    "TaskResumeUsecase": "vibe3.services.task.resume",
+    "TaskService": "vibe3.services.task.service",
+    "TaskShowResult": "vibe3.services.task.show",
+    "TaskStatusBucket": "vibe3.services.task.classifier",
     "VerdictService": "vibe3.services.verdict_service",
     # Functions
     "analyze_critical_files": "vibe3.services.pr.analysis",
@@ -231,7 +231,7 @@ _SYMBOL_MODULES = {
     "calculate_pr_risk_score": "vibe3.services.pr.analysis",
     "check_ref_exists": "vibe3.services.path_helpers",
     "classify_flow": "vibe3.services.flow_classifier",
-    "classify_task_issues_for_rendering": "vibe3.services.task_status_service",
+    "classify_task_issues_for_rendering": "vibe3.services.task.status",
     "create_flow_manager": "vibe3.services.flow_factory",
     "emit_issue_failed": "vibe3.services.event_helpers",
     "fail_executor_issue": "vibe3.services.issue.failure",
@@ -240,7 +240,7 @@ _SYMBOL_MODULES = {
     "fail_reviewer_issue": "vibe3.services.issue.failure",
     "format_agent_actor": "vibe3.services.actor_support",
     "filter_critical_files": "vibe3.services.pr.analysis",
-    "fetch_task_status_data": "vibe3.services.task_status_service",
+    "fetch_task_status_data": "vibe3.services.task.status",
     "format_issue_runtime_line": "vibe3.services.orchestra_status_service",
     "format_issue_summary_line": "vibe3.services.orchestra_status_service",
     "generate_score_report": "vibe3.services.pr.scoring",
