+++ src/vibe3/services/task/show.py
