+++ tests/vibe3/commands/test_task_show.py
@@ -7,7 +7,7 @@ from typer.testing import CliRunner
 
 from vibe3.cli import app
 from vibe3.models.flow import FlowStatusResponse
-from vibe3.services.task_service import (
+from vibe3.services.task import (
     TaskCommentSummary,
     TaskPRSummary,
     TaskRefSummary,
