+++ tests/vibe3/commands/test_task_status_dashboard.py
@@ -44,8 +44,8 @@ def _make_config_mock() -> MagicMock:
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_splits_assignee_ready_and_anomaly(
     mock_status_service_cls,
     mock_flow_service_cls,
@@ -140,8 +140,8 @@ def test_task_status_splits_assignee_ready_and_anomaly(
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_shows_flows_with_prs(
     mock_status_service_cls,
     mock_flow_service_cls,
@@ -195,8 +195,8 @@ def test_task_status_shows_flows_with_prs(
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_excludes_system_status_sections(
     mock_status_service_cls,
     mock_flow_service_cls,
@@ -240,8 +240,8 @@ def test_task_status_excludes_system_status_sections(
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_shows_missing_state_label_section(
     mock_status_service_cls,
     mock_flow_service_cls,
@@ -319,8 +319,8 @@ def test_task_status_shows_missing_state_label_section(
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_shows_active_exception_for_missing_assignee(
     mock_status_service_cls: MagicMock,
     mock_flow_service_cls: MagicMock,
@@ -379,8 +379,8 @@ def test_task_status_shows_active_exception_for_missing_assignee(
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_shows_governed_anomaly_section(
     mock_status_service_cls: MagicMock,
     mock_flow_service_cls: MagicMock,
@@ -506,8 +506,8 @@ class TestComputeEffectiveServerRunning:
 @patch(
     "vibe3.services.orchestra_status_service.OrchestraStatusService.fetch_live_snapshot"
 )
-@patch("vibe3.services.task_status_service.FlowService")
-@patch("vibe3.services.task_status_service.StatusQueryService")
+@patch("vibe3.services.task.status.FlowService")
+@patch("vibe3.services.task.status.StatusQueryService")
 def test_task_status_shows_remote_tasks_section(
     mock_status_service_cls,
     mock_flow_service_cls,
