+++ tests/vibe3/services/test_flow_block_mixin.py
@@ -14,7 +14,7 @@ def test_block_flow_calls_blocked_state_service() -> None:
         patch.object(service.store, "get_issue_links") as mock_get_links,
         patch.object(service.store, "update_flow_state"),
         patch.object(service.store, "add_event"),
-        patch("vibe3.services.task_service.TaskService"),
+        patch("vibe3.services.task.service.TaskService"),
         patch(
             "vibe3.services.blocked_state_service.BlockedStateService"
         ) as mock_blocked_service_cls,
